requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.13
django-crispy-forms==1.14.0
django-discover-runner==1.0
django-environ==0.9.0
django-resized==1.0.1
django-role-permissions==3.1.1
django-tinymce==3.4.0
gunicorn==20.1.0
html2text==2020.1.16
idna==3.3
numpy==1.21.6
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.5
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
uuid==1.30
